% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_from_ggplot.R
\name{as_highchart.ggplot}
\alias{as_highchart.ggplot}
\title{Convert a ggplot to a highchart (S3 method)}
\usage{
\method{as_highchart}{ggplot}(x, ...)
}
\arguments{
\item{x}{A ggplot object}

\item{...}{Additional arguments passed to \code{\link[=hc_from_ggplot]{hc_from_ggplot()}}}
}
\value{
A highcharter widget
}
\description{
S3 method allowing \code{as_highchart(p)} syntax for ggplot objects.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
as_highchart(p)
}

}
