% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_from_ggplot.R
\name{as_highchart}
\alias{as_highchart}
\title{Generic for as_highchart}
\usage{
as_highchart(x, ...)
}
\arguments{
\item{x}{An object to convert to a highchart}

\item{...}{Additional arguments}
}
\value{
A highcharter widget
}
\description{
Generic for as_highchart
}
